restored create_app; moved flask(__name__) to app init
